requirements.txt
biopython
keras
numpy
itertools